tests/array_2d/test_search_sequence.py
from src.array_2d.search_sequence import exist


def testExist():
    assert (
        exist(
            [
                ["A", "B", "C", "E"],
                ["S", "F", "C", "S"],
                ["A", "D", "E", "E"],
            ],
            "ABCCED",
        )
        == True
    )
    assert (
        exist(
            [
                ["A", "B", "C", "E"],
                ["S", "F", "C", "S"],
                ["A", "D", "E", "E"],
            ],
            "SEE",
        )
        == True
    )
    assert (
        exist(
            [
                ["A", "B", "C", "E"],
                ["S", "F", "C", "S"],
                ["A", "D", "E", "E"],
            ],
            "ABCB",
        )
        == False
    )
